Replace the old DistinctCount(Raw)ThetaSketchAggregationFunction with the new one
Make `DistinctCountRawThetaSketchAggregationFunction` the same usage as `DistinctCountThetaSketchAggregationFunction` and returns base64 encoded sketch.
Also fix the issue in BaseBrokerRequestHandler.updateColumnNames()
