+++ pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/AggregationFunctionFactory.java
@@ -137,10 +137,6 @@ public class AggregationFunctionFactory {
             return new DistinctCountThetaSketchAggregationFunction(arguments);
           case DISTINCTCOUNTRAWTHETASKETCH:
             return new DistinctCountRawThetaSketchAggregationFunction(arguments);
-          case THETASKETCH:
-            return new ThetaSketchAggregationFunction(arguments);
-          case RAWTHETASKETCH:
-            return new RawThetaSketchAggregationFunction(arguments);
           case IDSET:
             return new IdSetAggregationFunction(arguments);
           case COUNTMV:
